fix decomp bugs across 14 files — 5 more functions to 100%, 9 improved
Functions reaching 100%:
- HolmesClientClose (wrong field mBuffer→mRequestor, chained <<)
- PoseFatalities::CheckMatchingPose (&& → if/if/return eliminates bool mask)
- MetaPerformer::SetDefaultSongCharacter (removed nullSym, moved decls)
- SongStatusMgr::GetBestStars (loop bug: DifficultyOneHarder once not N)
- FlowRun::ResolveTarget (missing DirLoader proxy path)

Partially improved:
- ProfileMgr::IsContentUnlocked 88.5→99.9% (missing IsUnlockableAsset guard)
- Trie::inc_count 87.2→95.0% (unsigned char → unsigned int)
- Trie::dec_count 87.2→95.0% (same)
- NetCacheMgr::SetState 88.4→93.6% (MILO_FAIL workaround + enum names)
- MemInit 86.8→87.7% (reverse loop, free(mem) leak, static init)
- MCContainerXbox::Mount 87.9→88.4% (DWORD cast for unsigned compare)
- BinkMovieSys::Init (C linkage, non-virtual PlatformInit, assert string)
- op8 ByteGrinder 93.1→93.8% (from prior batch)

Correctness fixes without match% change:
- RndConsole::MoveLevel (FormatString workaround)
- PoseFatalities::InFatality (readability)

+++ src/system/os/Memcard_Xbox.cpp
@@ -170,7 +170,7 @@ MCContainerXbox::MCContainerXbox(const ContainerId &c) : MCContainer(c) {}
 
 MCResult MCContainerXbox::Mount(CreateType ct) {
     MILO_ASSERT(!IsMounted(), 0x144);
-    MILO_ASSERT(Cid().mDeviceId != XCONTENTDEVICE_ANY, 0x145);
+    MILO_ASSERT((DWORD)Cid().mDeviceId != XCONTENTDEVICE_ANY, 0x145);
     mDriveName = TheMC.GenerateDriveName(Cid().mDeviceId, Cid().mUserIndex);
     XCONTENT_DATA data;
     data.DeviceID = Cid().mDeviceId;
